+++ packages/locale-detector/rollup.config.js
@@ -9,22 +9,22 @@ import { terser } from 'rollup-plugin-terser'
 
 const getPath = (file) => path.resolve(__dirname, file)
 
-const files = glob(path.resolve(__dirname, `./src/**/*.ts`))
+const files = glob(path.resolve(__dirname, `./src/**/*.*ts`))
 	.map((file) =>
 		path
 			.resolve(file)
 			.substring(path.resolve(__dirname, `./src/`).length + 1)
 			.replace(/\\/g, '/'),
 	)
-	.filter((file) => !file.split('/').some((part) => part.startsWith('_') || part === 'index.ts'))
+	.filter((file) => !file.split('/').some((part) => part.startsWith('_')))
 
-const getOutputFormat = (file) => (file.startsWith('detectors/browser') || file.endsWith('.esm.ts') ? 'esm' : 'cjs')
+const getOutputFormat = (file) => (file.startsWith('detectors/browser') || file.endsWith('.mts') ? 'esm' : 'cjs')
 
 const config = files.map((file) => ({
 	input: getPath(`./src/${file}`),
 	output: [
 		{
-			file: getPath(`../../detectors/${file.replace('.ts', '.js')}`),
+			file: getPath(`../../detectors/${file.replace('.ts', '.js').replace('.mts', '.mjs')}`),
 			format: getOutputFormat(file),
 		},
 	],